refactor(schema): re-arrange conversion between arrow2
