+++ src/common/storage/src/parquet_rs.rs
@@ -18,8 +18,8 @@ use std::sync::Arc;
 use arrow_schema::Schema as ArrowSchema;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
+use databend_common_expression::converts::arrow::EXTENSION_KEY;
 use databend_common_expression::FieldIndex;
-use databend_common_expression::EXTENSION_KEY;
 use opendal::Operator;
 use parquet::arrow::parquet_to_arrow_schema;
 use parquet::file::footer::decode_footer;
