+++ src/query/expression/src/converts/arrow2/mod.rs
@@ -12,6 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod array;
-mod record_batch;
-mod schema;
+mod from;
+mod to;
+
+pub const ARROW_EXT_TYPE_EMPTY_ARRAY: &str = "EmptyArray";
+pub const ARROW_EXT_TYPE_EMPTY_MAP: &str = "EmptyMap";
+pub const ARROW_EXT_TYPE_VARIANT: &str = "Variant";
+pub const ARROW_EXT_TYPE_BITMAP: &str = "Bitmap";
+
+pub use to::set_validities;
