+++ src/query/expression/src/values.rs
@@ -22,15 +22,9 @@ use base64::engine::general_purpose;
 use base64::prelude::*;
 use borsh::BorshDeserialize;
 use borsh::BorshSerialize;
-use databend_common_arrow::arrow::bitmap::and;
 use databend_common_arrow::arrow::bitmap::Bitmap;
 use databend_common_arrow::arrow::bitmap::MutableBitmap;
 use databend_common_arrow::arrow::buffer::Buffer;
-use databend_common_arrow::arrow::compute::cast as arrow_cast;
-use databend_common_arrow::arrow::datatypes::DataType as ArrowDataType;
-use databend_common_arrow::arrow::datatypes::Field as ArrowField;
-use databend_common_arrow::arrow::datatypes::TimeUnit;
-use databend_common_arrow::arrow::offset::OffsetsBuffer;
 use databend_common_arrow::arrow::trusted_len::TrustedLen;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
@@ -740,12 +734,6 @@ impl PartialEq for Column {
     }
 }
 
-pub const EXTENSION_KEY: &str = "Extension";
-pub const ARROW_EXT_TYPE_EMPTY_ARRAY: &str = "EmptyArray";
-pub const ARROW_EXT_TYPE_EMPTY_MAP: &str = "EmptyMap";
-pub const ARROW_EXT_TYPE_VARIANT: &str = "Variant";
-pub const ARROW_EXT_TYPE_BITMAP: &str = "Bitmap";
-
 impl Column {
     pub fn len(&self) -> usize {
         match self {
@@ -994,296 +982,6 @@ impl Column {
         }
     }
 
-    pub fn arrow_field(&self) -> ArrowField {
-        let dummy = "DUMMY".to_string();
-        let is_nullable = matches!(&self, Column::Nullable(_));
-        let arrow_type: ArrowDataType = (&self.data_type()).into();
-        ArrowField::new(dummy, arrow_type, is_nullable)
-    }
-
-    pub fn as_arrow(&self) -> Box<dyn databend_common_arrow::arrow::array::Array> {
-        let arrow_type = self.arrow_field().data_type().clone();
-        match self {
-            Column::Null { len } => Box::new(
-                databend_common_arrow::arrow::array::NullArray::new_null(arrow_type, *len),
-            ),
-            Column::EmptyArray { len } => Box::new(
-                databend_common_arrow::arrow::array::NullArray::new_null(arrow_type, *len),
-            ),
-            Column::EmptyMap { len } => Box::new(
-                databend_common_arrow::arrow::array::NullArray::new_null(arrow_type, *len),
-            ),
-            Column::Number(NumberColumn::UInt8(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<u8>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::UInt16(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<u16>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::UInt32(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<u32>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::UInt64(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<u64>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::Int8(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i8>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::Int16(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i16>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::Int32(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i32>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::Int64(col)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i64>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Number(NumberColumn::Float32(col)) => {
-                let values =
-                    unsafe { std::mem::transmute::<Buffer<F32>, Buffer<f32>>(col.clone()) };
-                Box::new(
-                    databend_common_arrow::arrow::array::PrimitiveArray::<f32>::try_new(
-                        arrow_type, values, None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Number(NumberColumn::Float64(col)) => {
-                let values =
-                    unsafe { std::mem::transmute::<Buffer<F64>, Buffer<f64>>(col.clone()) };
-                Box::new(
-                    databend_common_arrow::arrow::array::PrimitiveArray::<f64>::try_new(
-                        arrow_type, values, None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Decimal(DecimalColumn::Decimal128(col, _)) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i128>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Decimal(DecimalColumn::Decimal256(col, _)) => {
-                let values = unsafe { std::mem::transmute(col.clone()) };
-                Box::new(
-                    databend_common_arrow::arrow::array::PrimitiveArray::<
-                        databend_common_arrow::arrow::types::i256,
-                    >::try_new(arrow_type, values, None)
-                    .unwrap(),
-                )
-            }
-            Column::Boolean(col) => Box::new(
-                databend_common_arrow::arrow::array::BooleanArray::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Binary(col) => {
-                let offsets: Buffer<i64> =
-                    col.offsets().iter().map(|offset| *offset as i64).collect();
-                Box::new(
-                    databend_common_arrow::arrow::array::BinaryArray::<i64>::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        col.data().clone(),
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::String(col) => {
-                let offsets: Buffer<i64> =
-                    col.offsets().iter().map(|offset| *offset as i64).collect();
-                Box::new(
-                    databend_common_arrow::arrow::array::BinaryArray::<i64>::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        col.data().clone(),
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Timestamp(col) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i64>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Date(col) => Box::new(
-                databend_common_arrow::arrow::array::PrimitiveArray::<i32>::try_new(
-                    arrow_type,
-                    col.clone(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Array(col) => {
-                let offsets: Buffer<i64> =
-                    col.offsets.iter().map(|offset| *offset as i64).collect();
-                Box::new(
-                    databend_common_arrow::arrow::array::ListArray::<i64>::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        col.values.as_arrow(),
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Map(col) => {
-                let offsets: Buffer<i32> =
-                    col.offsets.iter().map(|offset| *offset as i32).collect();
-                let values = match (&arrow_type, &col.values) {
-                    (ArrowDataType::Map(inner_field, _), Column::Tuple(fields)) => {
-                        let inner_type = inner_field.data_type.clone();
-                        Box::new(
-                            databend_common_arrow::arrow::array::StructArray::try_new(
-                                inner_type,
-                                fields.iter().map(|field| field.as_arrow()).collect(),
-                                None,
-                            )
-                            .unwrap(),
-                        )
-                    }
-                    (_, _) => unreachable!(),
-                };
-                Box::new(
-                    databend_common_arrow::arrow::array::MapArray::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        values,
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Bitmap(col) => {
-                let offsets: Buffer<i64> =
-                    col.offsets().iter().map(|offset| *offset as i64).collect();
-                Box::new(
-                    databend_common_arrow::arrow::array::BinaryArray::<i64>::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        col.data().clone(),
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-            Column::Nullable(col) => {
-                let arrow_array = col.column.as_arrow();
-                Self::set_validity(arrow_array.clone(), &col.validity)
-            }
-            Column::Tuple(fields) => Box::new(
-                databend_common_arrow::arrow::array::StructArray::try_new(
-                    arrow_type,
-                    fields.iter().map(|field| field.as_arrow()).collect(),
-                    None,
-                )
-                .unwrap(),
-            ),
-            Column::Variant(col) => {
-                let offsets: Buffer<i64> =
-                    col.offsets().iter().map(|offset| *offset as i64).collect();
-                Box::new(
-                    databend_common_arrow::arrow::array::BinaryArray::<i64>::try_new(
-                        arrow_type,
-                        unsafe { OffsetsBuffer::new_unchecked(offsets) },
-                        col.data().clone(),
-                        None,
-                    )
-                    .unwrap(),
-                )
-            }
-        }
-    }
-
-    pub fn set_validity(
-        arrow_array: Box<dyn databend_common_arrow::arrow::array::Array>,
-        validity: &Bitmap,
-    ) -> Box<dyn databend_common_arrow::arrow::array::Array> {
-        // merge Struct validity with the inner fields validity
-        let validity = match arrow_array.validity() {
-            Some(inner_validity) => and(inner_validity, validity),
-            None => validity.clone(),
-        };
-
-        match arrow_array.data_type() {
-            ArrowDataType::Null => arrow_array.clone(),
-            ArrowDataType::Extension(_, t, _) if **t == ArrowDataType::Null => arrow_array.clone(),
-            ArrowDataType::Struct(_) => {
-                let struct_array = arrow_array
-                    .as_any()
-                    .downcast_ref::<databend_common_arrow::arrow::array::StructArray>()
-                    .expect("fail to read from arrow: array should be `StructArray`");
-                let fields = struct_array
-                    .values()
-                    .iter()
-                    .map(|array| {
-                        let array = Self::set_validity(array.clone(), &validity);
-                        array.clone()
-                    })
-                    .collect::<Vec<_>>();
-                Box::new(
-                    databend_common_arrow::arrow::array::StructArray::try_new(
-                        arrow_array.data_type().clone(),
-                        fields,
-                        Some(validity),
-                    )
-                    .unwrap(),
-                )
-            }
-            _ => arrow_array.with_validity(Some(validity)),
-        }
-    }
-
     pub fn check_valid(&self) -> Result<()> {
         match self {
             Column::String(x) => x.check_valid(),
@@ -1312,469 +1010,6 @@ impl Column {
             }
             _ => Ok(()),
         }
-    }
-
-    pub fn from_arrow(
-        arrow_col: &dyn databend_common_arrow::arrow::array::Array,
-        data_type: &DataType,
-    ) -> Column {
-        fn from_arrow_by_array_type(
-            arrow_col: &dyn databend_common_arrow::arrow::array::Array,
-            array_type: &ArrowDataType,
-            data_type: &DataType,
-        ) -> Column {
-            let is_nullable = data_type.is_nullable();
-            let data_type = data_type.remove_nullable();
-            let column = match array_type {
-                ArrowDataType::Null => match data_type {
-                    DataType::EmptyArray => Column::EmptyArray {
-                        len: arrow_col.len(),
-                    },
-                    DataType::EmptyMap => Column::EmptyMap {
-                        len: arrow_col.len(),
-                    },
-                    _ => Column::Null {
-                        len: arrow_col.len(),
-                    },
-                },
-                ArrowDataType::Extension(name, _, _) if name == ARROW_EXT_TYPE_EMPTY_ARRAY => {
-                    Column::EmptyArray {
-                        len: arrow_col.len(),
-                    }
-                }
-                ArrowDataType::Extension(name, _, _) if name == ARROW_EXT_TYPE_EMPTY_MAP => {
-                    Column::EmptyMap {
-                        len: arrow_col.len(),
-                    }
-                }
-                ArrowDataType::UInt8 => Column::Number(NumberColumn::UInt8(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::UInt8Array>()
-                        .expect("fail to read from arrow: array should be `UInt8Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::UInt16 => Column::Number(NumberColumn::UInt16(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::UInt16Array>()
-                        .expect("fail to read from arrow: array should be `UInt16Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::UInt32 => Column::Number(NumberColumn::UInt32(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::UInt32Array>()
-                        .expect("fail to read from arrow: array should be `UInt32Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::UInt64 => Column::Number(NumberColumn::UInt64(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::UInt64Array>()
-                        .expect("fail to read from arrow: array should be `UInt64Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::Int8 => Column::Number(NumberColumn::Int8(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int8Array>()
-                        .expect("fail to read from arrow: array should be `Int8Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::Int16 => Column::Number(NumberColumn::Int16(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int16Array>()
-                        .expect("fail to read from arrow: array should be `Int16Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::Int32 => Column::Number(NumberColumn::Int32(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int32Array>()
-                        .expect("fail to read from arrow: array should be `Int32Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::Int64 => Column::Number(NumberColumn::Int64(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int64Array>()
-                        .expect("fail to read from arrow: array should be `Int64Array`")
-                        .values()
-                        .clone(),
-                )),
-                ArrowDataType::Float32 => {
-                    let col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Float32Array>()
-                        .expect("fail to read from arrow: array should be `Float32Array`")
-                        .values()
-                        .clone();
-                    let col = unsafe { std::mem::transmute::<Buffer<f32>, Buffer<F32>>(col) };
-                    Column::Number(NumberColumn::Float32(col))
-                }
-                ArrowDataType::Float64 => {
-                    let col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Float64Array>()
-                        .expect("fail to read from arrow: array should be `Float64Array`")
-                        .values()
-                        .clone();
-                    let col = unsafe { std::mem::transmute::<Buffer<f64>, Buffer<F64>>(col) };
-                    Column::Number(NumberColumn::Float64(col))
-                }
-                ArrowDataType::Boolean => Column::Boolean(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::BooleanArray>()
-                        .expect("fail to read from arrow: array should be `BooleanArray`")
-                        .values()
-                        .clone(),
-                ),
-                ArrowDataType::LargeBinary => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i64>>()
-                        .expect("fail to read from arrow: array should be `BinaryArray<i64>`");
-                    let offsets = arrow_col.offsets().clone().into_inner();
-
-                    let offsets =
-                        unsafe { std::mem::transmute::<Buffer<i64>, Buffer<u64>>(offsets) };
-                    // LargeBinary may be Extension data type variant and bitmap
-                    match data_type {
-                        DataType::Variant => {
-                            Column::Variant(StringColumn::new(arrow_col.values().clone(), offsets))
-                        }
-                        DataType::Bitmap => {
-                            Column::Bitmap(StringColumn::new(arrow_col.values().clone(), offsets))
-                        }
-                        _ => Column::String(StringColumn::new(arrow_col.values().clone(), offsets)),
-                    }
-                }
-                // TODO: deprecate it and use LargeBinary instead
-                ArrowDataType::Binary => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i32>>()
-                        .expect("fail to read from arrow: array should be `BinaryArray<i32>`");
-                    let offsets = arrow_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-                    // Binary may be Extension data type variant and bitmap
-                    match data_type {
-                        DataType::Variant => Column::Variant(StringColumn::new(
-                            arrow_col.values().clone(),
-                            offsets.into(),
-                        )),
-                        DataType::Bitmap => Column::Bitmap(StringColumn::new(
-                            arrow_col.values().clone(),
-                            offsets.into(),
-                        )),
-                        _ => Column::String(StringColumn::new(
-                            arrow_col.values().clone(),
-                            offsets.into(),
-                        )),
-                    }
-                }
-
-                ArrowDataType::FixedSizeBinary(size) => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::FixedSizeBinaryArray>()
-                        .expect("fail to read from arrow: array should be `FixedSizeBinaryArray`");
-
-                    let offsets = (0..arrow_col.len() as u64 + 1)
-                        .map(|x| x * (*size) as u64)
-                        .collect::<Vec<_>>();
-
-                    Column::String(StringColumn::new(
-                        arrow_col.values().clone(),
-                        offsets.into(),
-                    ))
-                }
-
-                // TODO: deprecate it and use LargeBinary instead
-                ArrowDataType::Utf8 => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Utf8Array<i32>>()
-                        .expect("fail to read from arrow: array should be `Utf8Array<i32>`");
-                    let offsets = arrow_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-
-                    Column::String(StringColumn::new(
-                        arrow_col.values().clone(),
-                        offsets.into(),
-                    ))
-                }
-                // TODO: deprecate it and use LargeBinary instead
-                ArrowDataType::LargeUtf8 => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Utf8Array<i64>>()
-                        .expect("fail to read from arrow: array should be `Utf8Array<i64>`");
-                    let offsets = arrow_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-                    Column::String(StringColumn::new(
-                        arrow_col.values().clone(),
-                        offsets.into(),
-                    ))
-                }
-
-                ArrowDataType::Timestamp(uint, _) => {
-                    let values = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int64Array>()
-                        .expect("fail to read from arrow: array should be `Int64Array`")
-                        .values();
-                    let convert = match uint {
-                        TimeUnit::Second => (1_000_000, 1),
-                        TimeUnit::Millisecond => (1_000, 1),
-                        TimeUnit::Microsecond => (1, 1),
-                        TimeUnit::Nanosecond => (1, 1_000),
-                    };
-
-                    let values = if convert.0 == 1 && convert.1 == 1 {
-                        values.clone()
-                    } else {
-                        let values = values
-                            .iter()
-                            .map(|x| x * convert.0 / convert.1)
-                            .collect::<Vec<_>>();
-                        values.into()
-                    };
-                    Column::Timestamp(values)
-                }
-                ArrowDataType::Date32 => Column::Date(
-                    arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::Int32Array>()
-                        .expect("fail to read from arrow: array should be `Int32Array`")
-                        .values()
-                        .clone(),
-                ),
-                ArrowDataType::Extension(name, box ty, None) if name == ARROW_EXT_TYPE_VARIANT => {
-                    match ty {
-                        ArrowDataType::LargeBinary => {
-                            let arrow_col = arrow_col
-                                .as_any()
-                                .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i64>>()
-                                .expect(
-                                    "fail to read from arrow: array should be `BinaryArray<i64>`",
-                                );
-                            let offsets = arrow_col.offsets().clone().into_inner();
-
-                            let offsets =
-                                unsafe { std::mem::transmute::<Buffer<i64>, Buffer<u64>>(offsets) };
-
-                            Column::Variant(StringColumn::new(arrow_col.values().clone(), offsets))
-                        }
-                        ArrowDataType::Binary => {
-                            let arrow_col = arrow_col
-                                .as_any()
-                                .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i32>>()
-                                .expect(
-                                    "fail to read from arrow: array should be `BinaryArray<i32>`",
-                                );
-                            let offsets = arrow_col
-                                .offsets()
-                                .buffer()
-                                .iter()
-                                .map(|x| *x as u64)
-                                .collect::<Vec<_>>();
-                            Column::Variant(StringColumn::new(
-                                arrow_col.values().clone(),
-                                offsets.into(),
-                            ))
-                        }
-                        _ => unreachable!(
-                            "fail to read from arrow: array should be `BinaryArray<i32>` or `BinaryArray<i64>`"
-                        ),
-                    }
-                }
-                ArrowDataType::List(f) => {
-                    let array_list = arrow_cast::cast(
-                        arrow_col,
-                        &ArrowDataType::LargeList(f.clone()),
-                        arrow_cast::CastOptions {
-                            wrapped: true,
-                            partial: true,
-                        },
-                    )
-                    .expect("list to large list cast should be ok");
-
-                    let values_col = array_list
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::ListArray<i64>>()
-                        .expect("fail to read from arrow: array should be `ListArray<i64>`");
-
-                    let array_type = data_type.as_array().unwrap();
-                    let values = Column::from_arrow(&**values_col.values(), array_type.as_ref());
-                    let offsets = values_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-                    Column::Array(Box::new(ArrayColumn {
-                        values,
-                        offsets: offsets.into(),
-                    }))
-                }
-                ArrowDataType::LargeList(_) => {
-                    let values_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::ListArray<i64>>()
-                        .expect("fail to read from arrow: array should be `ListArray<i64>`");
-
-                    let array_type = data_type.as_array().unwrap();
-                    let values = Column::from_arrow(&**values_col.values(), array_type.as_ref());
-                    let offsets = values_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-                    Column::Array(Box::new(ArrayColumn {
-                        values,
-                        offsets: offsets.into(),
-                    }))
-                }
-                ArrowDataType::Map(_, _) => {
-                    let map_type = data_type.as_map().unwrap();
-                    let map_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::MapArray>()
-                        .expect("fail to read from arrow: array should be `MapArray`");
-
-                    let values = Column::from_arrow(&**map_col.field(), map_type.as_ref());
-                    let offsets = map_col
-                        .offsets()
-                        .buffer()
-                        .iter()
-                        .map(|x| *x as u64)
-                        .collect::<Vec<_>>();
-                    Column::Map(Box::new(ArrayColumn {
-                        values,
-                        offsets: offsets.into(),
-                    }))
-                }
-                ArrowDataType::Struct(_) => {
-                    let struct_type = data_type.as_tuple().unwrap();
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::StructArray>()
-                        .expect("fail to read from arrow: array should be `StructArray`");
-                    let fields = arrow_col
-                        .values()
-                        .iter()
-                        .zip(struct_type.iter())
-                        .map(|(field, dt)| Column::from_arrow(&**field, dt))
-                        .collect::<Vec<_>>();
-                    Column::Tuple(fields)
-                }
-                ArrowDataType::Decimal(precision, scale) => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::PrimitiveArray<i128>>()
-                        .expect("fail to read from arrow: array should be `DecimalArray`");
-                    Column::Decimal(DecimalColumn::Decimal128(
-                        arrow_col.values().clone(),
-                        DecimalSize {
-                            precision: *precision as u8,
-                            scale: *scale as u8,
-                        },
-                    ))
-                }
-                ArrowDataType::Decimal256(precision, scale) => {
-                    let arrow_col = arrow_col
-                        .as_any()
-                        .downcast_ref::<databend_common_arrow::arrow::array::PrimitiveArray<
-                            databend_common_arrow::arrow::types::i256,
-                        >>()
-                        .expect("fail to read from arrow: array should be `DecimalArray`");
-
-                    let values = unsafe { std::mem::transmute(arrow_col.values().clone()) };
-                    Column::Decimal(DecimalColumn::Decimal256(values, DecimalSize {
-                        precision: *precision as u8,
-                        scale: *scale as u8,
-                    }))
-                }
-                ArrowDataType::Extension(name, box ty, None) if name == ARROW_EXT_TYPE_BITMAP => {
-                    match ty {
-                        ArrowDataType::LargeBinary => {
-                            let arrow_col = arrow_col
-                                .as_any()
-                                .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i64>>()
-                                .expect(
-                                    "fail to read from arrow: array should be `BinaryArray<i64>`",
-                                );
-                            let offsets = arrow_col.offsets().clone().into_inner();
-
-                            let offsets =
-                                unsafe { std::mem::transmute::<Buffer<i64>, Buffer<u64>>(offsets) };
-                            Column::Bitmap(StringColumn::new(arrow_col.values().clone(), offsets))
-                        }
-                        ArrowDataType::Binary => {
-                            let arrow_col = arrow_col
-                                .as_any()
-                                .downcast_ref::<databend_common_arrow::arrow::array::BinaryArray<i32>>()
-                                .expect(
-                                    "fail to read from arrow: array should be `BinaryArray<i32>`",
-                                );
-                            let offsets = arrow_col
-                                .offsets()
-                                .buffer()
-                                .iter()
-                                .map(|x| *x as u64)
-                                .collect::<Vec<_>>();
-                            Column::Bitmap(StringColumn::new(
-                                arrow_col.values().clone(),
-                                offsets.into(),
-                            ))
-                        }
-                        _ => unreachable!(
-                            "fail to read from arrow: array should be `BinaryArray<i32>` or `BinaryArray<i64>`"
-                        ),
-                    }
-                }
-                ArrowDataType::Extension(_name, box ty, _) => {
-                    from_arrow_by_array_type(arrow_col, ty, &data_type)
-                }
-                ty => unimplemented!("unsupported arrow type {ty:?}"),
-            };
-
-            if is_nullable {
-                let validity = arrow_col
-                    .validity()
-                    .cloned()
-                    .unwrap_or_else(|| Bitmap::new_constant(true, arrow_col.len()));
-                Column::Nullable(Box::new(NullableColumn { column, validity }))
-            } else {
-                column
-            }
-        }
-
-        from_arrow_by_array_type(arrow_col, arrow_col.data_type(), data_type)
     }
 
     pub fn random(ty: &DataType, len: usize) -> Self {
