+++ src/query/pipeline/sources/src/input_formats/impls/input_format_parquet.rs
@@ -336,8 +336,8 @@ impl BlockBuilderTrait for ParquetBlockBuilder {
             let fields: Vec<DataField> = rg
                 .fields_to_read
                 .iter()
-                .map(DataField::from)
-                .collect::<Vec<_>>();
+                .map(DataField::try_from)
+                .collect::<Result<Vec<_>>>()?;
 
             let input_schema = DataSchema::new(fields);
             let block = DataBlock::from_arrow_chunk(&chunk, &input_schema)?;
