+++ src/query/service/src/table_functions/infer_schema/infer_schema_table.rs
@@ -244,7 +244,7 @@ impl AsyncSource for InferSchemaSource {
                 if use_parquet2 {
                     let arrow_schema =
                         read_parquet_schema_async(&operator, &first_file.path).await?;
-                    TableSchema::from(&arrow_schema)
+                    TableSchema::try_from(&arrow_schema)?
                 } else {
                     let arrow_schema = read_parquet_schema_async_rs(
                         &operator,
