+++ src/query/storages/parquet/src/parquet2/statistics.rs
@@ -23,7 +23,7 @@ use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::types::DataType;
 use databend_common_expression::Column;
-use databend_common_expression::TableDataType;
+use databend_common_expression::DataField;
 use databend_common_storage::ColumnNodes;
 use databend_storages_common_table_meta::meta::ColumnStatistics;
 use databend_storages_common_table_meta::meta::StatisticsOfColumns;
@@ -44,12 +44,11 @@ pub fn collect_row_group_stats(
     // `column_nodes` is parallel to the schema, so we can iterate `column_nodes` directly.
     for (index, column_node) in column_nodes.column_nodes.iter().enumerate() {
         let field = &column_node.field;
-        let table_type: TableDataType = field.into();
-        let data_type = (&table_type).into();
+        let data_field = DataField::try_from(field).unwrap();
         let column_stats = pread::statistics::deserialize(field, rgs)?;
         stats_of_row_groups.insert(
             index,
-            BatchStatistics::from_statistics(&column_stats, &data_type)?,
+            BatchStatistics::from_statistics(&column_stats, data_field.data_type())?,
         );
     }
 
@@ -76,10 +75,9 @@ pub fn collect_basic_column_stats(
     // `column_nodes` is parallel to the schema, so we can iterate `column_nodes` directly.
     for column_node in column_nodes.column_nodes.iter() {
         let field = &column_node.field;
-        let table_type: TableDataType = field.into();
-        let data_type = (&table_type).into();
+        let data_field = DataField::try_from(field).unwrap();
         let column_stats = pread::statistics::deserialize(field, rgs)?;
-        let batch_stats = BatchStatistics::from_statistics(&column_stats, &data_type)?;
+        let batch_stats = BatchStatistics::from_statistics(&column_stats, data_field.data_type())?;
         let mut col_stats: BasicColumnStatistics = batch_stats.get(0).into();
         for rg_idx in 1..rgs.len() {
             col_stats.merge(batch_stats.get(rg_idx).into());
